choi :지윤님이 말씀하신 오류해결해드렸습니다.
member setting error
