--- src/main/java/Store/Store.java
@@ -1,32 +1,0 @@
-package Store;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Column;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.JoinColumn;
-import lombok.Getter;
-import lombok.Setter;
-import Menu.Menu;
-
-@Entity
-@Getter
-@Setter
-public class Store {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long storeId;
-
-    @Column(nullable = false)
-    private String storeName;
-
-    @Column(nullable = false)
-    private String storePhoneNumber;
-
-    @Column(nullable = false)
-    private String category;
-
-}
