+++ src/main/java/com/cl/youngri/Member/Member.java
@@ -1,0 +1,27 @@
+package com.cl.youngri.Member;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Column;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+public class Member {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long memberId;
+
+    @Column(unique = true, nullable = false)
+    private String username;
+
+    private String password;
+
+    private String number;
+}
+
