+++ src/main/java/com/cl/youngri/Member/MemberRepositroy.java
@@ -1,4 +1,4 @@
-package Member;
+package com.cl.youngri.Member;
 
 import org.springframework.data.jpa.repository.JpaRepository;
