+++ src/main/java/com/cl/youngri/Member/MemberService.java
@@ -1,12 +1,8 @@
-package Member;
+package com.cl.youngri.Member;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.lang.reflect.Member;
 
 @RequiredArgsConstructor
 @Service
