--- src/main/java/org/example/hackathon_2024/Hackathon2024Application.java
@@ -1,13 +1,0 @@
-package org.example.hackathon_2024;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Hackathon2024Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Hackathon2024Application.class, args);
-    }
-
-}
