--- src/test/java/org/example/hackathon_2024/Hackathon2024ApplicationTests.java
@@ -1,13 +1,0 @@
-package org.example.hackathon_2024;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Hackathon2024ApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
